+++ finality-provider/service/app.go
@@ -49,6 +49,7 @@ type FinalityProviderApp struct {
 	metrics *metrics.FpMetrics
 
 	createFinalityProviderRequestChan chan *CreateFinalityProviderRequest
+	unjailFinalityProviderRequestChan chan *UnjailFinalityProviderRequest
 	criticalErrChan                   chan *CriticalError
 }
 
@@ -114,6 +115,7 @@ func NewFinalityProviderApp(
 		eotsManager:                       em,
 		metrics:                           fpMetrics,
 		quit:                              make(chan struct{}),
+		unjailFinalityProviderRequestChan: make(chan *UnjailFinalityProviderRequest),
 		createFinalityProviderRequestChan: make(chan *CreateFinalityProviderRequest),
 		criticalErrChan:                   make(chan *CriticalError),
 	}, nil
@@ -251,12 +253,13 @@ func (app *FinalityProviderApp) Start() error {
 	app.startOnce.Do(func() {
 		app.logger.Info("Starting FinalityProviderApp")
 
-		app.wg.Add(5)
+		app.wg.Add(6)
 		go app.syncChainFpStatusLoop()
-		go app.registrationLoop()
 		go app.metricsUpdateLoop()
 		go app.monitorCriticalErr()
 		go app.monitorStatusUpdate()
+		go app.registrationLoop()
+		go app.unjailFpLoop()
 	})
 
 	return startErr
@@ -374,34 +377,39 @@ func (app *FinalityProviderApp) CreateFinalityProvider(
 }
 
 // UnjailFinalityProvider sends a transaction to unjail a finality-provider
-func (app *FinalityProviderApp) UnjailFinalityProvider(fpPk *bbntypes.BIP340PubKey) (string, error) {
-	_, err := app.fps.GetFinalityProvider(fpPk.MustToBTCPK())
-	if err != nil {
-		return "", fmt.Errorf("failed to get finality provider from db: %w", err)
+func (app *FinalityProviderApp) UnjailFinalityProvider(fpPk *bbntypes.BIP340PubKey) (*UnjailFinalityProviderResponse, error) {
+	// send request to the loop to avoid blocking the main thread
+	request := &UnjailFinalityProviderRequest{
+		btcPubKey:       fpPk,
+		errResponse:     make(chan error, 1),
+		successResponse: make(chan *UnjailFinalityProviderResponse, 1),
 	}
 
-	// Send unjail transaction
-	res, err := app.cc.UnjailFinalityProvider(fpPk.MustToBTCPK())
-	if err != nil {
-		return "", fmt.Errorf("failed to send unjail transaction: %w", err)
+	app.unjailFinalityProviderRequestChan <- request
+
+	select {
+	case err := <-request.errResponse:
+		return nil, err
+	case successResponse := <-request.successResponse:
+		_, err := app.fps.GetFinalityProvider(fpPk.MustToBTCPK())
+		if err != nil {
+			return nil, fmt.Errorf("failed to get finality provider from db: %w", err)
+		}
+
+		// Update finality-provider status in the local store
+		// set it to INACTIVE for now and it will be updated to
+		// ACTIVE if the fp has voting power
+		err = app.fps.SetFpStatus(fpPk.MustToBTCPK(), proto.FinalityProviderStatus_INACTIVE)
+		if err != nil {
+			return nil, fmt.Errorf("failed to update finality-provider status after unjailing: %w", err)
+		}
+
+		app.metrics.RecordFpStatus(fpPk.MarshalHex(), proto.FinalityProviderStatus_INACTIVE)
+
+		return successResponse, nil
+	case <-app.quit:
+		return nil, fmt.Errorf("finality-provider app is shutting down")
 	}
-
-	// Update finality-provider status in the local store
-	// set it to INACTIVE for now and it will be updated to
-	// ACTIVE if the fp has voting power
-	err = app.fps.SetFpStatus(fpPk.MustToBTCPK(), proto.FinalityProviderStatus_INACTIVE)
-	if err != nil {
-		return "", fmt.Errorf("failed to update finality-provider status after unjailing: %w", err)
-	}
-
-	app.metrics.RecordFpStatus(fpPk.MarshalHex(), proto.FinalityProviderStatus_INACTIVE)
-
-	app.logger.Info("successfully unjailed finality-provider",
-		zap.String("btc_pk", fpPk.MarshalHex()),
-		zap.String("txHash", res.TxHash),
-	)
-
-	return res.TxHash, nil
 }
 
 func (app *FinalityProviderApp) CreatePop(fpAddress sdk.AccAddress, fpPk *bbntypes.BIP340PubKey, passphrase string) (*bstypes.ProofOfPossessionBTC, error) {